src/main/java/lotto/model/BonusNumber.java
package lotto.model;

public class BonusNumber {
    public static final int MINIMUM_LOTTO_NUMBER = 1;
    public static final int MAXIMUM_LOTTO_NUMBER = 45;
    private int number;

    public BonusNumber(int number) {
        validateNumberRange(number);
        this.number = number;
    }

    private void validateNumberRange(int number) {
        if (number < MINIMUM_LOTTO_NUMBER || number > MAXIMUM_LOTTO_NUMBER) {
            throw new IllegalArgumentException("[ERROR] 보너스 숫자는 1이상 45이하의 숫자여야 합니다.");
        }
    }

    public int getNumber() {
        return number;
    }
}
